+++ src/tests/distributions.cc
@@ -12,21 +12,21 @@
 
 using namespace Smash;
 
-static float test_breit_wigner(const float srts, const float m, const float w) {
-  const float s = srts*srts;
-  const float sw2 = s * w * w;
-  const float smm = s - m * m;
+static double test_breit_wigner(const double srts, const double m, const double w) {
+  const double s = srts*srts;
+  const double sw2 = s * w * w;
+  const double smm = s - m * m;
   return 2.*s*w / (M_PI * (smm * smm + sw2));
 }
 
 TEST(breitwigner) {
   // tests the Breit-Wigner implementation for values between 0.05 and 10 GeV
   for (int s_i = 1; s_i < 200; ++s_i) {
-    float s = s_i * 0.05;
+    double s = s_i * 0.05;
     for (int m_i = 1; m_i < 200; ++m_i) {
-      float m = m_i * 0.05;
+      double m = m_i * 0.05;
       for (int w_i = 1; w_i < 200; ++w_i) {
-        float w = w_i * 0.05;
+        double w = w_i * 0.05;
         COMPARE_ABSOLUTE_ERROR(breit_wigner(s, m, w),
                                test_breit_wigner(s, m, w), 1e-6)
                 << s << "/" << m << "/" << w;
@@ -36,9 +36,9 @@ TEST(breitwigner) {
 }
 
 TEST(cauchy) {
-  const float m0 = 0.770;
-  const float gamma = 0.150;
-  const float peak_value = 1.f/(M_PI*gamma);
+  const double m0 = 0.770;
+  const double gamma = 0.150;
+  const double peak_value = 1.f/(M_PI*gamma);
   // cauchy: half maximum at full width
   FUZZY_COMPARE         (cauchy(m0,       m0, gamma), peak_value);
   COMPARE_ABSOLUTE_ERROR(cauchy(m0+gamma, m0, gamma), peak_value/2.f, 1e-6);
